refactor: move fiat_currency to lib/data/enums
