+++ lib/data/enums/fiat_currency.dart
